+++ examples/resources/nxos_default_qos_class_map_dscp/import-by-identity.tf
@@ -1,7 +1,0 @@
-import {
-  to = nxos_default_qos_class_map_dscp.example
-  identity = {
-    "class_map_name" : "<class_map_name>"
-    "value" : "<value>"
-  }
-}