+++ examples/resources/nxos_default_qos_class_map_dscp/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_default_qos_class_map_dscp.example
-  id = "<class_map_name>,<value>"
-}