+++ examples/resources/nxos_default_qos_class_map_dscp/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_default_qos_class_map_dscp.example "<class_map_name>,<value>"